+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Add setting to not load movie metadata
 * Input editor: pasting while selecting a range repeats the paste frames in range
 * Allow games to use the native Steam API
+* Implement SDL_GetWindowFromID() (#409)
 
 ### Changed